Modifications to finetune experiment

+++ src/test_bench.py
@@ -41,15 +41,18 @@ def predict(model, test_data):
 
 def log_predictions(y_predicted, model_name, path):
     model_file = model_name + '-res.csv'
-    f = open(path+model_file, "w+")
-    for i in xrange(len(y_predicted)):
-        line = '{0}, {1}, {2}, {3}\r\n'.format(str(i),
-                                               str(aa.confidence(y_predicted[i])),
-                                               str(np.argmax(y_predicted[i])),
-                                               str(y_predicted[i]))
-        f.write(line)
-    f.close()
-    print('Predictions for ' + model_file + ' written.')
+    if os.path.isfile(path + model_file):
+        print('Predictions for ' + model_file + ' already written!')
+    else:
+        f = open(path+model_file, "w+")
+        for i in xrange(len(y_predicted)):
+            line = '{0}, {1}, {2}, {3}\r\n'.format(str(i),
+                                                   str(aa.confidence(y_predicted[i])),
+                                                   str(np.argmax(y_predicted[i])),
+                                                   str(y_predicted[i]))
+            f.write(line)
+        f.close()
+        print('Predictions for ' + model_file + ' written.')
 
 
 def cifar_test():
@@ -118,7 +121,7 @@ def finetune_test():
         cdc_finetune.normalize()
         cdc_finetune.plot_cube(save=True, title=model_name0 + '-ft_selection', path=res_path)
 
-        print(finetune_data_args)
+        # print(finetune_data_args)
         dselec = np.concatenate((train_data_orig[0][:training_data_len],
                               np.array(operator.itemgetter(*finetune_data_args)(ft_data_orig[0]))))
         dlabels = np.concatenate((train_data_orig[1][:training_data_len],
@@ -133,14 +136,14 @@ def finetune_test():
 
         assert len(ft_data_selected) == 2 and len(ft_data_selected[0]) == 30000
 
-        model1, model_name1 = mt.fine_tune(model0, model_name0, ft_data_selected, val_data, 20, False, 'exp2', path=res_path)
+        model1, model_name1 = mt.fine_tune(model0, model_name0, ft_data_selected, val_data, 50, False, 'exp', path=res_path)
         y_predicted = predict(model1, formatted_test_data)
         log_predictions(y_predicted, model_name1, path=res_path)
         predicted_classes = np.argmax(y_predicted, axis=1)
         true_classes = np.argmax(formatted_test_data[1], axis=1)
         aa.accuracy(predicted_classes, true_classes)
 
-        model2, model_name2 = mt.fine_tune(model0, model_name0, train_data_ref, val_data, 20, False, 'ref2', path=res_path)
+        model2, model_name2 = mt.fine_tune(model0, model_name0, train_data_ref, val_data, 50, False, 'ref', path=res_path)
         y_predicted = predict(model2, formatted_test_data)
         log_predictions(y_predicted, model_name2, path=res_path)
         predicted_classes = np.argmax(y_predicted, axis=1)
